backend/requirements.txt
alembic==1.14.0
Flask-Cors==5.0.0
openai==1.59.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
SQLAlchemy==2.0.37
